pyproject.toml
[project]
name = "quantitative-system"
version = "0.1.0"
description = "AI驱动的量化交易系统 - 基于LLM的因子挖掘和股票筛选"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Quant Team", email = "houdd@mail.ustc.edu.cn"}
]
keywords = ["quantitative", "trading", "ai", "llm", "factor", "backtest"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyarrow>=14.0.0",
    "tushare>=1.3.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "typing-extensions>=4.8.0",
    "deepagents>=0.4.0,<0.5.0",
    "langchain-mcp-adapters>=0.2.0,<0.3.0",
    "langchain-openai>=0.2.0",
    "mcp>=1.26.0,<2.0.0",
    "exchange-calendars>=1.15.0",
    # Stock Asking MCP (workspace dependency)
    "stock-asking-mcp",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]
telemetry = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
]

[project.urls]
Homepage = "https://github.com/quantteam/quantitative-system"
Documentation = "https://quantitative-system.readthedocs.io"
Repository = "https://github.com/quantteam/quantitative-system"
Issues = "https://github.com/quantteam/quantitative-system/issues"

[project.scripts]
quant-factor = "factor_backtest_system.run_factor_mining:main"
quant-query = "stock_asking_system.run_stock_query:main"
quant-query-deepagent = "core.agent.run:main"
quant-mcp-server = "core.mcp.factor_tools_mcp:main"
datahub = "datahub.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# UV Workspace 配置
[tool.uv.workspace]
members = ["stock-asking-mcp"]

[tool.uv.sources]
stock-asking-mcp = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["config", "core", "datahub", "factor_backtest_system", "stock_asking_system", "core.agent"]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = [
    "factor_backtest_system/factor_scripts",
    "stock_asking_system/asking_scripts",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise without from (many existing except blocks)
    "C901",  # too complex
    "SIM108",  # prefer if/else over ternary for readability
]
per-file-ignores = { "__init__.py" = ["F401"], "datahub/entries.py" = ["N802"], "config/data_fields.py" = ["N815"], "config/factor_backtest_config.py" = ["N802"], "stock_asking_system/backtest/run_script_backtest.py" = ["E402"], "stock_asking_system/run_stock_query.py" = ["E402"], "factor_backtest_system/run_factor_mining.py" = ["E402"], "factor_backtest_system/backtest/factor_backtest.py" = ["E402"], "core/mcp/examples_usage.py" = ["E402"], "core/mcp/exceptions.py" = ["E402"], "core.agent/screening_scripts" = ["E402"], "core.agent/run.py" = ["E402"], "data2parquet/data_generator.py" = ["E402"], "data2parquet/data_saver.py" = ["E402"] }

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "tushare.*",
    "pandas.*",
    "numpy.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["core.agent.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["config", "core", "datahub", "factor_backtest_system", "stock_asking_system", "core.agent"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/screening_scripts/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
